fixup! combine location-list-lookup tests
