--- lldb/test/API/functionalities/location-list-lookup-cpp-member/Makefile
@@ -1,3 +1,0 @@
-CXX_SOURCES := main.cpp
-CFLAGS_EXTRAS := -O1
-include Makefile.rules
