+++ lldb/test/API/functionalities/location-list-lookup/Makefile
@@ -1,3 +1,3 @@
-C_SOURCES := main.c
+CXX_SOURCES := main.cpp
 CFLAGS_EXTRAS := -O1
 include Makefile.rules
