+++ lldb/test/API/functionalities/location-list-lookup/TestLocationListLookup.py
@@ -1,22 +1,15 @@
 ﻿"""Test that lldb picks the correct DWARF location list entry with a return-pc out of bounds."""
 
 import lldb
-from lldbsuite.test.decorators import *
 from lldbsuite.test.lldbtest import *
 from lldbsuite.test import lldbutil
 
 
-class LocationListLookupTestCase(TestBase):
-    def setUp(self):
-        # Call super's setUp().
-        TestBase.setUp(self)
-
-    @skipIf(oslist=["linux"], archs=["arm"])
-    def test_loclist(self):
+class CppMemberLocationListLookupTestCase(TestBase):
+    def test(self):
         self.build()
-        exe = self.getBuildArtifact("a.out")
 
-        # Create a target by the debugger.
+        exe = self.getBuildArtifact("a.out")
         target = self.dbg.CreateTarget(exe)
         self.assertTrue(target, VALID_TARGET)
         self.dbg.SetAsync(False)
@@ -27,12 +20,15 @@ class LocationListLookupTestCase(TestBase):
         self.assertTrue(process.IsValid())
         self.assertTrue(process.is_stopped)
 
-        # Find `main` on the stack, then
-        # find `argv` local variable, then
-        # check that we can read the c-string in argv[0]
+        # Find `bar` on the stack, then
+        # find `this` local variable, then
+        # check that we can read out the pointer value
         for f in process.GetSelectedThread().frames:
-            if f.GetDisplayFunctionName() == "main":
+            if f.GetDisplayFunctionName().startswith("Foo::bar"):
                 argv = f.GetValueForVariablePath("argv").GetChildAtIndex(0)
                 strm = lldb.SBStream()
                 argv.GetDescription(strm)
                 self.assertNotEqual(strm.GetData().find("a.out"), -1)
+
+                process.GetSelectedThread().SetSelectedFrame(f.idx)
+                self.expect_expr("this", result_type="Foo *")
